requirements.txt
asgiref==3.7.2
dj-rest-auth==5.1.0
Django==3.2.25
django-rest-auth==0.9.5
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.1
uritemplate==4.1.1
